[Cadencii/trunk] Cadencii_v3.3.4
git-svn-id: https://svn.sourceforge.jp/svnroot/cadencii/Cadencii/trunk@1434 9d680d4f-cbfa-477e-9905-73dbafb9b024

+++ __DRAFT__org.kbinani.vsq/UstVibrato.cs
@@ -72,7 +72,7 @@ namespace org.kbinani.vsq {
                 size = str.split( vec.get( spl, 1 ), spl, ",", false );
                 //VBR=65,180,70,20.0,17.6,82.8,49.8,100
                 if ( size >= 8 ) {
-                    Length = str.tof( spl[0] );
+                    Length = (float)str.tof( spl[0] );
                     Period = (float)str.tof( spl[1] );
                     Depth = (float)str.tof( spl[2] );
                     In = (float)str.tof( spl[3] );
